moved OtelMetrics to middleware package and added unit s to abnormal_terminations
